move loader tests to mirror directory
